requirements.txt
google-cloud-monitoring==0.33.0